% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_runrange.R
\name{addRunRange}
\alias{addRunRange}
\title{Add Run and Range values to a field experiment plots}
\usage{
addRunRange(sf_object, run1_range1)
}
\arguments{
\item{sf_object}{A sf object of geometry type POLYGON or MULTIPOLYGON}

\item{run1_range1}{Position of Run 1 and Range 1. Will take 4 different options. "TR","TL","BR","BL" denoting top right, top left, bottom right, bottom left respectively.}
}
\value{
A sf object with Run and Range columns added.
}
\description{
After generating plot boundaries, this functions allows users to assign Run and Range values. To identify what (where) is run1_range1 value,
assume you are standing on the **Southern end** of your **experiment width** side i.e. `exp_width` input in `corners_to_plot` function,
facing Northern end. Bottom is close to you, Top is away from you. Left on your left side and Right on your right side.
}
\examples{

dat <- read_input_dat(cornersData, 4326)

dat_plot <- corners_to_plots(sf_object = dat,
                      exp_length = 80, exp_width = 24.2,
                      n_runs = 11, n_ranges = 20)

datplot_w_RR <- addRunRange(dat_plot, "BL")

plot(datplot_w_RR)


}
