% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{corners_dat_csv}
\alias{corners_dat_csv}
\title{A data frame from a csv containing 4 XY coordinates and a id variable}
\format{
## `corners_dat_csv`
A data frame with 4 rows and 3 columns:
\describe{
  \item{X}{Longitude}
  \item{Y}{Latitude}
  \item{id}{id}
  ...
}
}
\usage{
corners_dat_csv
}
\description{
XY coordinates in CRS 4326 of a
}
\keyword{datasets}
