% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cornersData}
\alias{cornersData}
\title{A data frame containing 4 XY coordinates and a id variable}
\format{
A data frame with 4 rows and 3 columns:
\describe{
  \item{X}{Longitude}
  \item{Y}{Latitude}
  \item{id}{id}
  ...
}
}
\usage{
cornersData
}
\description{
XY coordinates in WGS84 EPSG:4326 of a field experiment in Australia. Data was originally recorded in a csv format.
}
\keyword{datasets}
