% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg_to_rad.R
\name{deg_2_rad}
\alias{deg_2_rad}
\title{Convert degrees to radians}
\usage{
deg_2_rad(degree_angle)
}
\arguments{
\item{degree_angle}{A numerical value. Angle in degrees.}
}
\value{
A numerical value degree in radians
}
\description{
Convert degrees to radians
}
