% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractPlotImages.R
\name{extract_plots_from_rasters}
\alias{extract_plots_from_rasters}
\title{Extract individual plots photos from raster}
\usage{
extract_plots_from_rasters(
  raster,
  plot_layout,
  plot_subset = NULL,
  dir_loc = NULL,
  file_type = ".png"
)
}
\arguments{
\item{raster}{A raster object}

\item{plot_layout}{A sf object containing polygons}

\item{plot_subset}{Subset of plots to be extracted rather than all plots}

\item{dir_loc}{A directory location to write files}

\item{file_type}{".png" or ".tif" type images to write}
}
\value{
progress message
}
\description{
Extract individual plots photos from raster
}
