% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_offsets.R
\name{find_offsets}
\alias{find_offsets}
\title{Find offsets to move plots horizontally and,or vertically}
\usage{
find_offsets(sf_object)
}
\arguments{
\item{sf_object}{A sf object of geometry POINT or POLYGON with Run and Range fields}
}
\value{
A matrix of X and Y offsets for horizontal and vertical shifts
}
\description{
Produces 1 plot unit offset values for moving the plots horizontally or vertically. If you
}
\examples{

dat <- corners_to_plots(cornersData, 80, 24.2, 11, 20)

dat_w_RR <- addRunRange(dat, "BL")

find_offsets(dat_w_RR)

}
