% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plot.R
\name{get_plot}
\alias{get_plot}
\title{Get raster for a specific plot or shape}
\usage{
get_plot(
  raster_obj,
  vector_obj,
  plot_no = NULL,
  RGB = NULL,
  stretchRGB = FALSE
)
}
\arguments{
\item{raster_obj}{Raster object as SpatRaster}

\item{vector_obj}{Shapefile as a SpatVector}

\item{plot_no}{OPTIONAL Plot number if it is a experimental layout. By default first shape is used.}

\item{RGB}{OPTIONAL, if TRUE, Raster will presented as a single layer RGB for better viz}

\item{stretchRGB}{OPTIONAL, if TRUE RGB values are stretched}
}
\value{
a raster object as SpatRaster
}
\description{
Get raster for a specific plot or shape
}
