% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vi.R
\name{get_vi}
\alias{get_vi}
\title{Get vegetation index from RGB raster}
\usage{
get_vi(raster_obj, which_vi = NULL)
}
\arguments{
\item{raster_obj}{RGB raster}

\item{which_vi}{Select vegetation index. EXI, VDVI, NGDRI, MGRVI}
}
\value{
raster object of vegetation index
}
\description{
Get vegetation index from RGB raster
}
