% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points_to_plots.R
\name{make_plots}
\alias{make_plots}
\title{Make plots from points i.e. plot centers or use to adjust length and the width of plots in an existing plot layout}
\usage{
make_plots(sf_obj, plot_length, plot_width)
}
\arguments{
\item{sf_obj}{A sf object with geometry type POINTS referring to plot center or POLYGONS}

\item{plot_length}{Length of individual plots}

\item{plot_width}{Width of individual plots}
}
\value{
A sf object of geometry type POLYGON with the same CRS as input sf_obj
}
\description{
Make plots from points i.e. plot centers or use to adjust length and the width of plots in an existing plot layout
}
\examples{

#reading a corner points data

dat <- read_input_dat(cornersData, 4326)

#Making a default plot grid using exp length and width, and using number of plots

dat_plots <- corners_to_plots(dat, 80, 24.2, 11, 20)

plot(dat_plots)

#Adjusting the size of the individual plots

adj_plots <- make_plots(dat_plots, 2.5, 1.2)

plot(adj_plots)


}
