% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_plots.R
\name{make_split_plots}
\alias{make_split_plots}
\title{Create split plot centers}
\usage{
make_split_plots(sf_obj, split_dist, split_into)
}
\arguments{
\item{sf_obj}{A sf object from a shape file, shp or geojson}

\item{split_dist}{Numeric input in meters indicating amount of offset to use in creating new plots.}

\item{split_into}{Integer input, number of new plots to generate from one plot, takes 1 or 2}
}
\value{
A sf object with geometry POINT
}
\description{
Splitting a single regular plot into two or three plots.
}
