% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moveAllPlots.R
\name{moveAllPlots}
\alias{moveAllPlots}
\title{Move all plots}
\usage{
moveAllPlots(sf_obj, xy_move)
}
\arguments{
\item{sf_obj}{sf object to type POLYGON to move.}

\item{xy_move}{a numeric vector of length 2. X and Y move distance in meters.}
}
\value{
A sf object
}
\description{
Move all the plots in your experiment to a new center.
}
