% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_to_4c.R
\name{multicorners_to_4corners}
\alias{multicorners_to_4corners}
\title{Multi-corner points to 4 corner points}
\usage{
multicorners_to_4corners(sf_object)
}
\arguments{
\item{sf_object}{A sf object or a data frame with three columns, geometry and
an id column to group by corners. The data frame needs to contain columns Longitude,
Latitude and id.}
}
\value{
sf object with 4 points
}
\description{
To increase the accuracy of finding the center point of an field experiment,
users can use this function where multiple points for each corner is collected.
This function then use all those points to find the average corner points and
improve the precision of the center.
}
