% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPlotR-package.R
\docType{package}
\name{plotPlotR-package}
\alias{plotPlotR}
\alias{plotPlotR-package}
\title{plotPlotR: Field experiment plot creation, segmentation and manipulation}
\description{
This package provides users with a set of tools to easily create and manipulate field experiment plot boundaries in R. The functions are based on the underpinning of "sf" package implementation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AgDataNomad/plotPlotR}
  \item \url{https://agdatanomad.github.io/plotPlotR/}
  \item Report bugs at \url{https://github.com/AgDataNomad/plotPlotR/issues}
}

}
\author{
\strong{Maintainer}: Aswin Singaram Natarajan \email{sin17m@csiro.au} (\href{https://orcid.org/0000-0001-8981-6353}{ORCID})

Authors:
\itemize{
  \item Todd Collins \email{col469@csiro.au} (\href{https://orcid.org/0000-0001-9286-0494}{ORCID}) [contributor]
}

}
\keyword{internal}
