% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points_to_plots.R
\name{points_to_plots}
\alias{points_to_plots}
\title{Make plots from points}
\usage{
points_to_plots(sf_obj, plot_length, plot_width)
}
\arguments{
\item{sf_obj}{A sf object with geometry type POLYGON referring to plot centers.}

\item{plot_length}{Length of individual plots}

\item{plot_width}{Width of individual plots}
}
\value{
A sf object of geometry type POLYGON with the same CRS as input sf_obj
}
\description{
Make plots from points
}
