% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rad_to_deg.R
\name{rad_2_deg}
\alias{rad_2_deg}
\title{radian to degree}
\usage{
rad_2_deg(rad_value)
}
\arguments{
\item{rad_value}{Rotation angle in radian.}
}
\value{
A numeric value of degree angle. Values will be between Zero and 360.
}
\description{
Convert radians to degree angle.
}
