% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rot_angle.R
\name{rot_angle}
\alias{rot_angle}
\title{Orientation of field experiment layout}
\usage{
rot_angle(sf_object)
}
\arguments{
\item{sf_object}{A sf object of geometry MULTIPOLYGON or POINT, or a data frame with 4 XY}
}
\value{
A numeric value of angle in radians
}
\description{
This function takes a sf object of geometry MULTIPOLYGON or POINT and find it rotation on it center.
For example, a perfect North facing field experiment has a rotation of Zero degrees,
East would be 90 degrees and so on. Make sure data in Longitude or X and Latitude or Y order
}
\examples{

rot_angle(cornersData)

}
