% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rot_plots.R
\name{rot_plots}
\alias{rot_plots}
\title{Rotate individual plots on its center}
\usage{
rot_plots(sf_object, rotation_angle)
}
\arguments{
\item{sf_object}{A MULTIPOLYGON sf object, representing plots on a field experiment.}

\item{rotation_angle}{Rotation angle in radians. Use function rot_angle() to find rotation angle or pass radians as numeric input.}
}
\value{
Returns a rotated sf object
}
\description{
Rotate individual plots on its center
}
