% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rot_shape.R
\name{rot_shape}
\alias{rot_shape}
\title{Rotate an entire sf object}
\usage{
rot_shape(rot_angle)
}
\arguments{
\item{rot_angle}{rotation angle in radians}
}
\value{
Rotated sf object
}
\description{
Rotate an entire sf object
}
