% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgs84_to_unitM.R
\name{wgs84_to_unitM}
\alias{wgs84_to_unitM}
\title{Converting coordinate to a unit M CRS}
\usage{
wgs84_to_unitM(sf_object, CRSinMeters)
}
\arguments{
\item{sf_object}{A sf object or a data frame that with X and Y columns containing Longitude and Latitude data}

\item{CRSinMeters}{CRS that is in unit M. Check epsg.io to find appropriate CRS for your location}
}
\value{
A sf with unit M CRS
}
\description{
Converting coordinate to a unit M CRS
}
\examples{

wgs84_to_unitM(corners_dat_csv, 28355)

}
