% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_output.R
\name{write_output}
\alias{write_output}
\title{Write outputs}
\usage{
write_output(output_dat, file_name)
}
\arguments{
\item{output_dat}{sf object to write to disk}

\item{file_name}{file name with file path and extension}
}
\description{
Write outputs as shapefile
}
