% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_output.R
\name{write_output_obj}
\alias{write_output_obj}
\title{Write outputs}
\usage{
write_output_obj(output_dat, file_name, overwrite = NULL)
}
\arguments{
\item{output_dat}{sf object to write to disk}

\item{file_name}{file name with file path and extension}

\item{overwrite}{OPTIONAL, TRUE to overwrite file}
}
\description{
Write outputs of \strong{plotPlotR} such as experimental layouts as shapefile; supports most shapefile formats
}
