% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_offsets.R
\name{find_offsets}
\alias{find_offsets}
\title{Find offsets to move plots horizontally and/or vertically}
\usage{
find_offsets(sf_object, runAlias, rangeAlias)
}
\arguments{
\item{sf_object}{A sf object of geometry POINT or POLYGON with Run and Range fields}

\item{runAlias}{Alias for Run in the sf object. Alias for Run could be row, run, Run, Row, ROW, RUN etc}

\item{rangeAlias}{Alias for Range in the sf object. Alias for Range could range, Range, col, Column column etc}
}
\value{
A matrix of X and Y offsets for horizontal and vertical moves
}
\description{
Calculates 1 meter offset values for moving the plots horizontally or vertically to account for any discrepencies during sowing.
}
\examples{

dat <- wgs84_to_unitM(cornersData, 28355) #28355 is a unit M CRS

dat_plots <- corners_to_plots(dat, 80, 24.2, 11, 20)

dat_w_RR <- addRunRange(dat_plots, "BL", 11, 20)

find_offsets(dat_w_RR, "Run", "Range")

}
