% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movePlotRanges.R
\name{movePlotRanges}
\alias{movePlotRanges}
\title{Move ranges in your field experiment horizontally or vertically}
\usage{
movePlotRanges(sf_obj, moveDirection, moveDistance, rangeSelection)
}
\arguments{
\item{sf_obj}{A sf object of geometry POLYGON or MULTIPOLYGON with Run and Range columns}

\item{moveDirection}{Direction of the move in reference Run 1 Range 1. Takes "up", "down", "left" or "right" as inputs.}

\item{moveDistance}{Numeric value in meters to move}

\item{rangeSelection}{Numeric value or numeric vector of a subset of ranges or all ranges}
}
\value{
a sf object of geometry POLYGON with Run and Range columns
}
\description{
Users can select a subset of ranges or all ranges
}
\details{
Move distance is in unit meters

Move direction is in relation to the position of Run 1 and Range 1

Move plots in your field experiment.

This is unlikely as tractor sowing affects entire Run and almost never a subset of Ranges.
}
\examples{

dat <- read_input_dat(cornersData, 4326, 28355)
dat_plots <- corners_to_plots(dat, 80, 24.2, 11, 20)
dat_plots_RR <- addRunRange(dat_plots, "BL")


moved_plots <- movePlotRanges(dat_plots_RR, "right", 1.2, c(1,3,9))

plot(moved_plots)


}
