% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movePlotRuns.R
\name{movePlotRuns}
\alias{movePlotRuns}
\title{Move runs in your field experiment horizontally or vertically}
\usage{
movePlotRuns(sf_obj, moveDirection, moveDistance, runSelection)
}
\arguments{
\item{sf_obj}{A sf object of geometry POLYGON or MULTIPOLYGON with Run and Range columns}

\item{moveDirection}{Direction of the move in reference to Run 1 Range 1. Takes "up", "down", "left" or "right" as inputs.}

\item{moveDistance}{Numeric value in meters to move}

\item{runSelection}{Numeric value or numeric vector of a subset of runs or all runs}
}
\description{
Run `find_offsets` before running `movePlotRuns`
}
\details{
Move plots in your field experiment
}
\examples{

dat <- wgs84_to_unitM(cornersData, 28355)
dat_plots <- corners_to_plots(dat, 80, 24.2, 11, 20)
dat_plots_RR <- addRunRange(dat_plots, "BL")

find_offsets(dat_plots_RR, "Run", "Range")

moved_plots <- movePlotRuns(dat_plots_RR, "up", 4, c(3,7))

plot(moved_plots)

}
