% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_input_dat.R
\name{read_input_dat}
\alias{read_input_dat}
\title{Read input data}
\usage{
read_input_dat(input_dat, input_CRS, to_CRS_M)
}
\arguments{
\item{input_dat}{CSV file containing XY or any shape file formats.}

\item{input_CRS}{Numeric input. CRS of input data as EPSG code}

\item{to_CRS_M}{NUmeric input. Corresponding CRS in unit M to input CRS as EPSG code}
}
\value{
A sf object with CRS
}
\description{
Read input data
}
\examples{
cornersData

dat <- read_input_dat(cornersData, 4326, 28355)

dat

}
